Refactor get_offers_by_requirements to offers
