+++ src/dstack/_internal/server/services/fleets.py
@@ -2,12 +2,14 @@ import random
 import string
 import uuid
 from datetime import timezone
-from typing import List, Optional, Union
+from typing import List, Optional, Tuple, Union
 
 from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import joinedload, selectinload
 
+from dstack._internal.core.backends import BACKENDS_WITH_CREATE_INSTANCE_SUPPORT
+from dstack._internal.core.backends.base import Backend
 from dstack._internal.core.errors import (
     ForbiddenError,
     ResourceExistsError,
@@ -24,8 +26,20 @@ from dstack._internal.core.models.fleets import (
     SSHHostParams,
     SSHParams,
 )
-from dstack._internal.core.models.instances import InstanceStatus, SSHKey
-from dstack._internal.core.models.profiles import SpotPolicy
+from dstack._internal.core.models.instances import (
+    DockerConfig,
+    InstanceConfiguration,
+    InstanceOfferWithAvailability,
+    InstanceStatus,
+    SSHKey,
+)
+from dstack._internal.core.models.pools import Instance
+from dstack._internal.core.models.profiles import (
+    DEFAULT_POOL_TERMINATION_IDLE_TIME,
+    Profile,
+    SpotPolicy,
+    TerminationPolicy,
+)
 from dstack._internal.core.models.resources import ResourcesSpec
 from dstack._internal.core.models.runs import Requirements, get_policy_map
 from dstack._internal.core.models.users import GlobalRole
@@ -38,12 +52,19 @@ from dstack._internal.server.models import (
     ProjectModel,
     UserModel,
 )
+from dstack._internal.server.services import offers as offers_services
 from dstack._internal.server.services import pools as pools_services
+from dstack._internal.server.services.docker import parse_image_name
+from dstack._internal.server.services.jobs.configurators.base import (
+    get_default_image,
+    get_default_python_verison,
+)
 from dstack._internal.server.services.locking import (
     get_locker,
     string_to_lock_id,
 )
 from dstack._internal.server.services.projects import get_member, get_member_permissions
+from dstack._internal.utils import common as common_utils
 from dstack._internal.utils import random_names
 from dstack._internal.utils.logging import get_logger
 from dstack._internal.utils.ssh import pkey_from_str
@@ -115,7 +136,6 @@ async def get_plan(
     spec: FleetSpec,
 ) -> FleetPlan:
     # TODO: refactor offers logic into a separate module to avoid depending on runs
-    from dstack._internal.server.services.runs import get_create_instance_offers
 
     offers = []
     if spec.configuration.ssh_config is None:
@@ -142,6 +162,40 @@ async def get_plan(
         max_offer_price=max((offer.price for offer in offers), default=None),
     )
     return plan
+
+
+async def get_create_instance_offers(
+    project: ProjectModel,
+    profile: Profile,
+    requirements: Requirements,
+    exclude_not_available=False,
+    fleet_model: Optional[FleetModel] = None,
+) -> List[Tuple[Backend, InstanceOfferWithAvailability]]:
+    multinode = False
+    master_job_provisioning_data = None
+    if fleet_model is not None:
+        fleet = fleet_model_to_fleet(fleet_model)
+        multinode = fleet.spec.configuration.placement == InstanceGroupPlacement.CLUSTER
+        for instance in fleet_model.instances:
+            jpd = pools_services.get_instance_provisioning_data(instance)
+            if jpd is not None:
+                master_job_provisioning_data = jpd
+                break
+
+    offers = await offers_services.get_offers_by_requirements(
+        project=project,
+        profile=profile,
+        requirements=requirements,
+        exclude_not_available=exclude_not_available,
+        multinode=multinode,
+        master_job_provisioning_data=master_job_provisioning_data,
+    )
+    offers = [
+        (backend, offer)
+        for backend, offer in offers
+        if backend.TYPE in BACKENDS_WITH_CREATE_INSTANCE_SUPPORT
+    ]
+    return offers
 
 
 async def create_fleet(
@@ -383,6 +437,91 @@ def is_fleet_in_use(fleet_model: FleetModel, instance_nums: Optional[List[int]] 
 def is_fleet_empty(fleet_model: FleetModel) -> bool:
     active_instances = [i for i in fleet_model.instances if not i.deleted]
     return len(active_instances) == 0
+
+
+async def create_instance(
+    session: AsyncSession,
+    project: ProjectModel,
+    user: UserModel,
+    profile: Profile,
+    requirements: Requirements,
+) -> Instance:
+    offers = await get_create_instance_offers(
+        project=project,
+        profile=profile,
+        requirements=requirements,
+        exclude_not_available=True,
+    )
+
+    # Raise error if no backends suppport create_instance
+    backend_types = set((backend.TYPE for backend, _ in offers))
+    if all(
+        (backend_type not in BACKENDS_WITH_CREATE_INSTANCE_SUPPORT)
+        for backend_type in backend_types
+    ):
+        backends = ", ".join(sorted(backend_types))
+        raise ServerClientError(
+            f"Backends {backends} do not support create_instance. Try to select other backends."
+        )
+
+    pool = await pools_services.get_or_create_pool_by_name(session, project, profile.pool_name)
+    instance_name = await pools_services.generate_instance_name(
+        session=session,
+        project=project,
+        pool_name=pool.name,
+    )
+
+    termination_policy = profile.termination_policy or TerminationPolicy.DESTROY_AFTER_IDLE
+    termination_idle_time = profile.termination_idle_time
+    if termination_idle_time is None:
+        termination_idle_time = DEFAULT_POOL_TERMINATION_IDLE_TIME
+
+    instance = InstanceModel(
+        id=uuid.uuid4(),
+        name=instance_name,
+        instance_num=0,
+        project=project,
+        pool=pool,
+        created_at=common_utils.get_current_datetime(),
+        status=InstanceStatus.PENDING,
+        unreachable=False,
+        profile=profile.json(),
+        requirements=requirements.json(),
+        instance_configuration=None,
+        termination_policy=termination_policy,
+        termination_idle_time=termination_idle_time,
+    )
+    logger.info(
+        "Added a new instance %s",
+        instance.name,
+        extra={
+            "instance_name": instance.name,
+            "instance_status": InstanceStatus.PENDING.value,
+        },
+    )
+    session.add(instance)
+    await session.commit()
+
+    project_ssh_key = SSHKey(
+        public=project.ssh_public_key.strip(),
+        private=project.ssh_private_key.strip(),
+    )
+    dstack_default_image = parse_image_name(get_default_image(get_default_python_verison()))
+    instance_config = InstanceConfiguration(
+        project_name=project.name,
+        instance_name=instance_name,
+        instance_id=str(instance.id),
+        ssh_keys=[project_ssh_key],
+        job_docker_config=DockerConfig(
+            image=dstack_default_image,
+            registry_auth=None,
+        ),
+        user=user.name,
+    )
+    instance.instance_configuration = instance_config.json()
+    await session.commit()
+
+    return pools_services.instance_model_to_instance(instance)
 
 
 def _check_can_manage_ssh_fleets(user: UserModel, project: ProjectModel):
